deployment/yolo-app/requirements.txt
# requirements.txt
streamlit
ultralytics
boto3
opencv-python-headless
numpy
Pillow
